refactor(script): use ts for cli-generate
